[REF] website_crm: include website_crm_phone_validation
Website crm phone validation does not hold code that cannot be located in
website_crm as dependencies are now the same. Let us merge those two
modules !

Linked to task 2033777
