+++ addons/website_crm/__manifest__.py
@@ -1,12 +1,16 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
 {
     'name': 'Contact Form',
     'category': 'Website',
     'sequence': 54,
     'summary': 'Generate leads from a contact form',
-    'version': '2.0',
+    'version': '2.1',
     'description': """
-Generate leads or opportunities in the CRM app from a contact form published on the Contact us page of your website. This form can be customized thanks to the *Form Builder* module (available in Odoo Enterprise).
-    """,
+Generate leads or opportunities in the CRM app from a contact form published on the Contact us page of your website.
+This form can be customized thanks to the *Form Builder* module (available in Odoo Enterprise).
+
+This module includes contact phone and mobile numbers validation.""",
     'depends': ['website_form', 'crm'],
     'data': [
         'data/website_crm_data.xml',
