--- addons/website_crm_phone_validation/__manifest__.py
@@ -1,22 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-{
-    'name': 'Contact Form Number Validation',
-    'summary': 'Validate and format contact form numbers',
-    'sequence': '9999',
-    'category': 'Hidden',
-    'description': """
-Contact Number Validation on Website
-====================================
-
-Validate contact (phone,mobile) numbers and normalize them on leads and contacts:
-- use the national format for your company country
-- use the international format for all others
-        """,
-    'data': [],
-    'depends': [
-        'website_crm',
-        'website_form',
-    ],
-    'auto_install': True,
-}
