--- addons/website_crm_phone_validation/controllers/__init__.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import website_form
